Move .from_array tests to separate file.
